requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==0.27.10
pydantic==1.10.11
